Workaround issue where it's not possible to query the bind-pose scale value in blender, causing issues importing scale keyframes.
The value is destroyed in blender after applying a scale to the EditBone. This is not how it works in maya, (or any other professional 3d suite), making it hard to compute absolute scale keyframes (the most common way to store them). This creates a custom blender property on PoseBone which stores the original bind pose scale value so we can use it to calculate the proper absolute scale keyframe later.